Adds a frame analyzer manager
The frame analyzer manager is a class responsible for keeping track
of the analyzers available in the application.

To do so, a decorator class FrameAnalyzerMetadata has been added.
This class also acts as an entity class that contains the metadata
of a frame analyzer.

When acting as a decorator, this class automatically creates an
instance of itself and registers it in the FrameAnalyzerManager
class, that can be queried to get a list of all the available frame
analyzers.