Improves frame analyzer API
Some functionalities have been added to the frame analyzer API.

Specifically, the FrameAnalyzerManager now allows retrieving the
metadata of FrameAnalyer instances and subclasses, and allows
loading all the classes in a module, so that the frame analyzers
are registered in the manager.

In addition, the CompositeFrameAnalyzer now allows changing the
child component.