Improves frame process events and communication with input and output
FrameProcess has been improved at several points. The first point is
that now it emits events when a frame is inputted and also when it is
outputted. The latter avoids the usage of composite frame outputs in
cases where the frame is not needed.

The second point is that now the frame output also receives the frame
number. This was not required by any particular case, but it seemed
coherent with the input and frame processor implementations.

The third point is that now when the inputted frame is None the
frame processor will finish. This is a way for frame inputs to
notify frame processor that the input is finished, avoiding active
processing when it is not needed.

The fourth point is that now CompositeFrameOutput allows removing
child outputs. In addition, this class is now thread safe.

Finally, the frame processor locks have been "relaxed", reducing the
code they cover.