Improves LIVIA properties type management
The management of the type in LIVIA property classes has been
improved and now the property type is inferred from the return
annotation, which now is required.

In addition, the FrameAnalyserPropertyMetadata now is a subclass of
LiviaPropertyMetadata and includes methods to get and set the
property value of an instance.