Fixes and improves frame processors
The base frame processor (FrameProcessor) has been improved to reduce
locking time. This change fixes one problem existing with frame
inputs with delays or locks (such as FileFrameInput), as the frame
processor was blocked while in internal locks.

On the other hand, the AsyncAnalyizerFrameProcessor now includes a
modification persistence option that allows to specify the number of
frames where a modification is still valid if no new modification is
created.