Improves async frame analyzer
The AsyncAnalyzerFrameProcessor has been improved for processing more
frames. To do so, the thread that waits for new frames now only is
blocked when no frame is available (i.e. checks if current frame is
None before blocking).

This increases the rate of frames that it processes. It is important
to note that we have found that the time required by the GPU for
processing a frame is lower when the frames are continuously
sumbitted. Therefore, this modification also has this positive side
effect.