Upgrades and completes dependencies
OpenCV dependency version has been upgraded and changed to the
headless version of the library, as GUI utils are not required.

Numpy was also included as a dependency. It was not previously
included as it was a transitive dependency of OpenCV, but, as it is
directly used in the code, it was included in the project's
configuration.

In addition, development dependencies have been added to setup.py and
requirements.txt file has been created.