Improves asynchronous analyzer frame processor
Due to asynchrony, two strange situations could occur in the current
implementation of the frame processor.

First, since a modification alters the ndarray of the frame it
receives, an analyzer may have to analyze an already modified frame.

Secondly, even if an analyzer can process frames immediately, these
modifications might not be displayed, since the step from saving the
frame in the shared variable to modifying the frame is immediate.

To solve the first problem, now the frame that is passed to the
analyzer is a copy of the original frame. This copy is never
modified.

To solve the second problem, a "delay" parameter has been added to
allow setting a waiting time between the storage of the frame in the
shared variable and its modification.