Improves object tracking
Object tracking has been improved in several aspects.

The tracked object class now has an unique identifier and
allows specifiying whether a frame is empty because it was not
processed or because the object was not found in the frame. It also
includes now a custom __str__ method that shows the detection status
of each object.

The ObjectLocation class now includes a method that adjust the object
to a box, for example, to adjust a location to the frame boundaries.

A minor bug in the DetectedObjectGroup was fixed. It was causing the
object construction to fail when an iterable without index access was
provided as parameter.

Finally, the ObjectTrackingAndClassifyingFrameModification class now
allows specifying the text color and adjust the boxes to the frame
boundaries.